+++ lib/zaniah.rb
@@ -35,6 +35,7 @@ require_relative "zaniah/accessibility"
 require_relative "zaniah/app"
 require_relative "zaniah/process_pool"
 require_relative "zaniah/input"
+require_relative "zaniah/drag_drop"
 require_relative "zaniah/element"
 Zaniah.autoload(:SVG, File.expand_path("zaniah/svg", __dir__))
 require_relative "zaniah/scroll_state"